vote-interface: move vote ix data types to new module
